+++ CHANGES.rst
@@ -1,4 +1,4 @@
-Changes to be released in next version
+Changes in 1.3.7 (2021-05-12)
 =================================================
 
 ✨ Features
@@ -23,6 +23,9 @@ Changes to be released in next version
 
 Others
  * 
+
+Improvements:
+ * Upgrade MatrixKit version ([v0.14.12](https://github.com/matrix-org/matrix-ios-kit/releases/tag/v0.14.12)).
 
 Changes in 1.3.6 (2021-05-07)
 =================================================